+++ example/src/Shared/ExampleComponent.tsx
@@ -1,5 +1,9 @@
 import React from 'react'
-import * as Tabs from 'react-native-collapsible-tab-view'
+import {
+  Tabs,
+  CollapsibleRef,
+  CollapsibleProps,
+} from 'react-native-collapsible-tab-view'
 
 import Albums from './Albums'
 import Article from './Article'
@@ -8,9 +12,9 @@ import { HEADER_HEIGHT } from './Header'
 
 type Props = {
   emptyContacts?: boolean
-} & Partial<Tabs.CollapsibleProps>
+} & Partial<CollapsibleProps>
 
-const Example = React.forwardRef<Tabs.CollapsibleRef, Props>(
+const Example = React.forwardRef<CollapsibleRef, Props>(
   ({ emptyContacts, ...props }, ref) => {
     return (
       <Tabs.Container ref={ref} headerHeight={HEADER_HEIGHT} {...props}>