+++ example/src/types.ts
@@ -4,4 +4,5 @@ export type ExampleComponentType<P = object> = React.FC<P> & {
   backgroundColor?: string
   statusBarStyle?: 'light-content' | 'dark-content'
   appbarElevation?: number
+  platform?: 'ios' | 'android'
 }